Rename gem to safe_migrations
Gem naming guidelines are to use underscores for multiple words.  This then
matches the file the user will require, making it easier for the user to start
using your gem.

Dashes should be reserved for extensions (adding function to another gem).
"This usually corresponds to a / in the require statement (and therefore your
gem’s directory structure) and a :: in the name of your main class or module."